Change to Homepage
This change makes it so that when a user goes to the site, if they have
a session already, they will automatically go to the My Programs
(welcome user) screen. if no session is started they are sent to the
home page where they can log in or read about the site and possibly
register.