WorkInProgess AddProgram
This is where I'm at up to this point. The code adds into the db and
works, I created new tables for the time being, since all the
information in the other tables isn't being used at the moment I wanted
to be able to get what we needed to work up for now. Next I will start
to work on the NextSong.php, it won't take long once I figure out how to
carry over the Program_ID variable from the AddProgramInitial.php. any
suggestions?